+++ doc/RELEASES.md
@@ -1,13 +1,15 @@
 Release notes
 =============
 
-FIXME
------
+2016.11.30
+----------
 
 - better detection and logging of unsupported Entity MIB
 - SNMPv1 support
 - SNMPv3 support (without SNMP contexts)
 - fix SNMP-GET and WALK call
+- added  logaction() to each API call (one line per action distinct log file)
+- added `clientinfo` parameter to a few commands to identify the upstream user
 
 25.11.2016
 ----------